Раздлены классы на отдельные модули, в интерфейсе теперь нет protected атрибутов
